requirements.txt
certifi==2019.11.28
chardet==3.0.4
docutils==0.16
fire==0.2.1
idna==2.8
lockfile==0.12.2
python-daemon==2.2.4
requests==2.22.0
six==1.14.0
termcolor==1.1.0
urllib3==1.25.7
XlsxWriter==1.2.7
